% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_portfolioMTM.R
\name{portfolioMTM}
\alias{portfolioMTM}
\title{portfolioMTM}
\usage{
portfolioMTM(x, shockpar = 0, shockInd = FALSE, output = "p")
}
\arguments{
\item{x}{portfolio input as tibble}

\item{shockpar}{parallel ytm shock applied to portfolio}

\item{shockInd}{individual shocks to ytm}

\item{output}{mtm of potfolio or individual bonds}
}
\value{
portfolio mark-to-market
}
\description{
computes the mark-to-market value of the portfolio
}
\author{
Philippe Cote
}
