% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcppPortParallel}
\alias{rcppPortParallel}
\title{Bond sensitivities}
\usage{
rcppPortParallel(x, stepSize)
}
\arguments{
\item{x}{bond details as matrix}

\item{stepSize}{sensitivities stepsize}
}
\value{
matrix of sensitivities
}
\description{
Bond sensitivities
}
