% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTLappBonds-package.R
\docType{package}
\name{RTLappBonds-package}
\alias{RTLappBonds}
\alias{RTLappBonds-package}
\title{RTLappBonds: Bond risk at portfolio level}
\description{
Develop your understanding of fixed income risk through active experimentation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/risktoollib/RTLappBonds}
  \item Report bugs at \url{https://github.com/risktoollib/RTLappBonds/issues}
}

}
\author{
\strong{Maintainer}: Philippe Cote \email{pcote@ualberta.ca}

}
\keyword{internal}
